requirements.txt

google-generativeai
sendgrid
python-docx
python-dotenv
markdown
